% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dde.R
\name{calculate_history}
\alias{calculate_history}
\title{Calculate history for DDE system}
\usage{
calculate_history(t, p)
}
\arguments{
\item{t}{temperature}

\item{p}{a \link{list}}
}
\description{
Calculate history for DDE system
}
