% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dde.R
\name{calculate_IC}
\alias{calculate_IC}
\title{Calculate initial conditions for DDE system}
\usage{
calculate_IC(A0, t0, pars)
}
\arguments{
\item{A0}{initial number of adults}

\item{t0}{initial temperature (assumed constant for \code{t<0})}

\item{pars}{a \link{list}}
}
\description{
Calculate initial conditions for DDE system
}
