% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_culex_deterministic}
\alias{create_culex_deterministic}
\title{Create deterministic culex model object}
\usage{
create_culex_deterministic(tau_E, tau_L, tau_P, dt, parameters)
}
\arguments{
\item{tau_E}{vector of egg delays}

\item{tau_L}{vector of larvae delays}

\item{tau_P}{vector of pupae delays}

\item{dt}{size of time step}

\item{parameters}{a \link{list} of parameters}

\item{p}{number of patches}
}
\description{
Create deterministic culex model object
}
