% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_culex_infection_stochastic}
\alias{create_culex_infection_stochastic}
\title{Create stochastic culex infection model object}
\usage{
create_culex_infection_stochastic(
  p,
  tau_E,
  tau_L,
  tau_P,
  tau_EIP,
  dt,
  psi,
  n_species
)
}
\arguments{
\item{p}{number of patches}

\item{tau_E}{vector of egg delays}

\item{tau_L}{vector of larvae delays}

\item{tau_P}{vector of pupae delays}

\item{tau_EIP}{vector of extrinsic incubation period delays}

\item{dt}{size of time step}

\item{psi}{adult movement matrix}

\item{n_species}{number of host species}
}
\description{
Create stochastic culex infection model object
}
