\name{culex-package}
\alias{culex-package}
\alias{culex}
\docType{package}
\title{
\packageTitle{culex}
}
\description{
\packageDescription{culex}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{culex}
\packageIndices{culex}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{culex}

Maintainer: \packageMaintainer{culex}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
