% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dde.R
\name{culex_dde}
\alias{culex_dde}
\title{DDE system}
\usage{
culex_dde(t, y, params)
}
\arguments{
\item{t}{time}

\item{y}{state}

\item{params}{a \link{list}}
}
\description{
R translation of culex model by David Ewing \href{https://github.com/davewi13/Temperate-Mosquito-DDE/blob/master/Chapter\%202\%20DDE\%20code.f90}{(code here)}
}
