% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{culex_parameters}
\alias{culex_parameters}
\title{Model parameters}
\usage{
culex_parameters(
  phi = 1.4,
  lambda = 6.3,
  mu = 10.3,
  gamma = 1.21,
  L = 51,
  max_egg = 200,
  q1 = 0.2024,
  q2 = 74.48,
  q3 = 0.2456,
  nu_0E = 0.0157,
  nu_1E = 20.5,
  nu_2E = 7,
  nu_0L = 0.0157,
  nu_1L = 20.5,
  nu_2L = 7,
  nu_0P = 0.0157,
  nu_1P = 20.5,
  nu_2P = 7,
  alpha_A = 2.166e-08,
  beta_A = 4.483,
  alpha_E = 0.0022,
  beta_E = 1.77,
  alpha_L = 0.00315,
  beta_L = 1.12,
  alpha_P = 0.0007109,
  beta_P = 1.8865648,
  a = 1,
  h = 0.002,
  r = 0.001,
  V = 200,
  death_max = 1,
  death_min_a = 0.01,
  gon_min = 0.0333,
  maturation_min = 0.016667,
  eip_q = 7.83e-05,
  eip_tmax = 45.2,
  eip_tmin = 11.4,
  eip_min = 0.005
)
}
\arguments{
\item{phi}{temperature phase}

\item{lambda}{temperature A}

\item{mu}{temperature M}

\item{gamma}{temperature power}

\item{L}{latitude for photoperiod}

\item{max_egg}{maximum egg raft size R}

\item{q1}{gonotrophic cycle KG}

\item{q2}{gonotrophic cycle QG}

\item{q3}{gonotrophic cycle BG}

\item{nu_0E}{egg stage mortality U3}

\item{nu_1E}{egg stage mortality U4}

\item{nu_2E}{egg stage mortality U5}

\item{nu_0L}{larval stage mortality U3}

\item{nu_1L}{larval stage mortality U4}

\item{nu_2L}{larval stage mortality U5}

\item{nu_0P}{pupal stage mortality U3}

\item{nu_1P}{pupal stage mortality U4}

\item{nu_2P}{pupal stage mortality U5}

\item{alpha_A}{adult stage mortality ALPHA}

\item{beta_A}{adult stage mortality BETA}

\item{alpha_E}{egg stage development ALPHA}

\item{beta_E}{egg stage development BETA}

\item{alpha_L}{larval stage development ALPHA}

\item{beta_L}{larval stage development BETA}

\item{alpha_P}{pupal stage development ALPHA}

\item{beta_P}{pupal stage development BETA}

\item{a}{larval predation}

\item{h}{larval predation}

\item{r}{larval predation}

\item{V}{larval predation}

\item{death_max}{maximum death rate for aquatic stages}

\item{death_min_a}{minimum death rate for adult stage}

\item{gon_min}{minimum rate of gonotrophic cycle}

\item{maturation_min}{minimum rate of aquatic stage development}

\item{eip_q}{EIP parameter}

\item{eip_tmax}{EIP parameter}

\item{eip_tmin}{EIP parameter}

\item{eip_min}{minimum rate of EIP progression}
}
\description{
Most of these parameters can be found at \link[=https://github.com/davewi13/Temperate-Mosquito-DDE/blob/master/Chapter 2 DDE code.f90]{https://github.com/davewi13/Temperate-Mosquito-DDE/blob/master/Chapter\%202\%20DDE\%20code.f90}.
}
