% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{death_adult_rate}
\alias{death_adult_rate}
\title{Adult mortality rate}
\usage{
death_adult_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a \link{list}}
}
\description{
Adult mortality rate
}
