% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{death_larvae_rate}
\alias{death_larvae_rate}
\title{Larvae mortality rate}
\usage{
death_larvae_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a \link{list}}
}
\description{
Larvae mortality rate
}
