% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{death_pupae_rate}
\alias{death_pupae_rate}
\title{Pupae mortality rate}
\usage{
death_pupae_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a \link{list}}
}
\description{
Pupae mortality rate
}
