% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delays.R
\name{egg_maturation_rate}
\alias{egg_maturation_rate}
\title{Egg maturation rate}
\usage{
egg_maturation_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a list}
}
\description{
Egg maturation rate
}
