% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{eip_rate}
\alias{eip_rate}
\title{Extrinsic incubation development rate}
\usage{
eip_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a list}
}
\description{
Extrinsic incubation development rate
}
