% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_A_deterministic}
\alias{get_A_deterministic}
\title{Return adults for deterministic culex model}
\usage{
get_A_deterministic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return adults for deterministic culex model
}
