% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_A_stochastic}
\alias{get_A_stochastic}
\title{Return adults for stochastic culex model}
\usage{
get_A_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return adults for stochastic culex model
}
