% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_AE_infection_stochastic}
\alias{get_AE_infection_stochastic}
\title{Get incubating adults for stochastic culex infection model}
\usage{
get_AE_infection_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Get incubating adults for stochastic culex infection model
}
