% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_E_stochastic}
\alias{get_E_stochastic}
\title{Return eggs for stochastic culex model}
\usage{
get_E_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return eggs for stochastic culex model
}
