% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_L_deterministic}
\alias{get_L_deterministic}
\title{Return larvae for deterministic culex model}
\usage{
get_L_deterministic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return larvae for deterministic culex model
}
