% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_L_infection_deterministic}
\alias{get_L_infection_deterministic}
\title{Get larvae for deterministic culex infection model}
\usage{
get_L_infection_deterministic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Get larvae for deterministic culex infection model
}
