% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_L_stochastic}
\alias{get_L_stochastic}
\title{Return larvae for stochastic culex model}
\usage{
get_L_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return larvae for stochastic culex model
}
