% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_P_deterministic}
\alias{get_P_deterministic}
\title{Return pupae for deterministic culex model}
\usage{
get_P_deterministic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return pupae for deterministic culex model
}
