% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_P_stochastic}
\alias{get_P_stochastic}
\title{Return pupae for stochastic culex model}
\usage{
get_P_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Return pupae for stochastic culex model
}
