% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_PI_culex_infection_stochastic}
\alias{get_PI_culex_infection_stochastic}
\title{Get infected pupae for stochastic culex infection model}
\usage{
get_PI_culex_infection_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Get infected pupae for stochastic culex infection model
}
