% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_q_infection_stochastic}
\alias{get_q_infection_stochastic}
\title{Get feeding habit for stochastic culex infection model}
\usage{
get_q_infection_stochastic(mod)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}
}
\description{
Get feeding habit for stochastic culex infection model
}
