% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{gonotrophic}
\alias{gonotrophic}
\title{Gonotrophic cycle length}
\usage{
gonotrophic(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a \link{list}}
}
\description{
Gonotrophic cycle length
}
