% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{larvae_maturation_rate}
\alias{larvae_maturation_rate}
\title{Larvae maturation rate}
\usage{
larvae_maturation_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a list}
}
\description{
Larvae maturation rate
}
