% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{oviposition}
\alias{oviposition}
\title{Egg laying rate}
\usage{
oviposition(d, G, pars)
}
\arguments{
\item{d}{diapause}

\item{G}{duration of gonotrophic cycle}

\item{pars}{a \link{list}}
}
\description{
Egg laying rate
}
