% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{pupae_maturation_rate}
\alias{pupae_maturation_rate}
\title{Pupae maturation rate}
\usage{
pupae_maturation_rate(temp, pars)
}
\arguments{
\item{temp}{temperature}

\item{pars}{a list}
}
\description{
Pupae maturation rate
}
