% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_A_deterministic}
\alias{set_A_deterministic}
\title{Set adults for deterministic culex model}
\usage{
set_A_deterministic(mod, A)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{A}{a vector}
}
\description{
Set adults for deterministic culex model
}
