% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_AE_culex_infection_stochastic}
\alias{set_AE_culex_infection_stochastic}
\title{Set incubating adults for stochastic culex infection model}
\usage{
set_AE_culex_infection_stochastic(mod, A)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{A}{a matrix}
}
\description{
Set incubating adults for stochastic culex infection model
}
