% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_AE_infection_deterministic}
\alias{set_AE_infection_deterministic}
\title{Set incubating adults for deterministic culex infection model}
\usage{
set_AE_infection_deterministic(mod, A)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{A}{a matrix}
}
\description{
Set incubating adults for deterministic culex infection model
}
