% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_AS_infection_stochastic}
\alias{set_AS_infection_stochastic}
\title{Set susceptible adults for stochastic culex infection model}
\usage{
set_AS_infection_stochastic(mod, A)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{A}{a row vector}
}
\description{
Set susceptible adults for stochastic culex infection model
}
