% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_E_culex_infection_stochastic}
\alias{set_E_culex_infection_stochastic}
\title{Set eggs for stochastic culex infection model}
\usage{
set_E_culex_infection_stochastic(mod, E)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{E}{a matrix}
}
\description{
Set eggs for stochastic culex infection model
}
