% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_E_infection_deterministic}
\alias{set_E_infection_deterministic}
\title{Set eggs for deterministic culex infection model}
\usage{
set_E_infection_deterministic(mod, E)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{E}{a matrix}
}
\description{
Set eggs for deterministic culex infection model
}
