% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_EI_culex_infection_stochastic}
\alias{set_EI_culex_infection_stochastic}
\title{Set infected eggs for stochastic culex infection model}
\usage{
set_EI_culex_infection_stochastic(mod, E)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{E}{a matrix}
}
\description{
Set infected eggs for stochastic culex infection model
}
