% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_f_culex_infection_stochastic}
\alias{set_f_culex_infection_stochastic}
\title{Set feeding rate for stochastic culex infection model}
\usage{
set_f_culex_infection_stochastic(mod, f)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{f}{a row vector (feeding rate by patch)}
}
\description{
Set feeding rate for stochastic culex infection model
}
