% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_kappa_infection_stochastic}
\alias{set_kappa_infection_stochastic}
\title{Set kappa for stochastic culex infection model}
\usage{
set_kappa_infection_stochastic(mod, kappa)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{kappa}{a matrix (each column indicates net infectiousness of each host species in that patch)}
}
\description{
Set kappa for stochastic culex infection model
}
