% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_L_culex_infection_stochastic}
\alias{set_L_culex_infection_stochastic}
\title{Set larvae for stochastic culex infection model}
\usage{
set_L_culex_infection_stochastic(mod, L)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{L}{a matrix}
}
\description{
Set larvae for stochastic culex infection model
}
