% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_LI_infection_deterministic}
\alias{set_LI_infection_deterministic}
\title{Set infected larvae for deterministic culex infection model}
\usage{
set_LI_infection_deterministic(mod, L)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{L}{a matrix}
}
\description{
Set infected larvae for deterministic culex infection model
}
