% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_P_infection_deterministic}
\alias{set_P_infection_deterministic}
\title{Set pupae for deterministic culex infection model}
\usage{
set_P_infection_deterministic(mod, P)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{P}{a matrix}
}
\description{
Set pupae for deterministic culex infection model
}
