% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_PI_culex_infection_stochastic}
\alias{set_PI_culex_infection_stochastic}
\title{Set infected pupae for stochastic culex infection model}
\usage{
set_PI_culex_infection_stochastic(mod, P)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{P}{a matrix}
}
\description{
Set infected pupae for stochastic culex infection model
}
