% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_q_infection_deterministic}
\alias{set_q_infection_deterministic}
\title{Set feeding habit for deterministic culex infection model}
\usage{
set_q_infection_deterministic(mod, q)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{q}{a matrix (columns must sum to \code{1}; each column indicates proportion of bites allocated to each host species in that patch)}
}
\description{
Set feeding habit for deterministic culex infection model
}
