% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{step_culex_infection_deterministic}
\alias{step_culex_infection_deterministic}
\title{Step deterministic culex infection model}
\usage{
step_culex_infection_deterministic(mod, parameters)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{parameters}{a \link{list} of parameters}
}
\description{
Step deterministic culex infection model
}
