% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{step_culex_stochastic}
\alias{step_culex_stochastic}
\title{Step stochastic culex model}
\usage{
step_culex_stochastic(mod, parameters)
}
\arguments{
\item{mod}{an \link[methods:BasicClasses]{methods::externalptr} object}

\item{parameters}{a \link{list} of parameters}
}
\description{
Step stochastic culex model
}
