% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tau_diffeqn}
\alias{tau_diffeqn}
\title{Differential equations describing change in delays}
\usage{
tau_diffeqn(t, y, params)
}
\arguments{
\item{t}{time}

\item{y}{state}

\item{params}{a list of parameters}
}
\description{
This function is meant to be passed to \link[deSolve:ode]{deSolve::ode}.
}
