% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_culex_infection_deterministic}
\alias{create_culex_infection_deterministic}
\title{Create deterministic culex infection model object}
\usage{
create_culex_infection_deterministic(
  tau_E,
  tau_L,
  tau_P,
  tau_EIP,
  dt,
  parameters,
  n_species
)
}
\arguments{
\item{tau_E}{vector of egg delays}

\item{tau_L}{vector of larvae delays}

\item{tau_P}{vector of pupae delays}

\item{tau_EIP}{vector of extrinsic incubation period delays}

\item{dt}{size of time step}

\item{parameters}{a \link{list} of parameters}

\item{n_species}{number of host species}
}
\description{
Create deterministic culex infection model object
}
